wifi: extract _GsmManager into its own module (#37911)
* wifi: extract _GsmManager into its own module

* gsm_manager: preserve comments/breadcrumbs from extracted code and move imports to module top
